requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
dotenv-python==0.0.1
load-dotenv==0.1.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
